+++ Src/Configuration.DkimSigner/GitHub/ApiWrapper.cs
@@ -14,8 +14,7 @@ namespace Configuration.DkimSigner.GitHub
         {
             List<Release> aoRelease = null;
 
-            string sJson = null;
-            sJson = Api.MakeRequest(Api.CreateRequest("/repos/pro/dkim-exchange/releases"));
+            string sJson = Api.MakeRequest(Api.CreateRequest("/repos/pro/dkim-exchange/releases"));
 
             if (sJson != null)
             {