+++ Src/Exchange.DkimSigner/Configuration/DomainElement.cs
@@ -1,24 +1,12 @@
-﻿using System;
-using System.IO;
-using System.Security.Cryptography;
-using DkimSigner.RSA;
+﻿using System.IO;
 
-namespace ConfigurationSettings
+namespace Exchange.DkimSigner.Configuration
 {
     public class DomainElement
     {
         public string Domain { get; set; }
         public string Selector { get; set; }
         public string PrivateKeyFile { get; set; }
-
-        /// <summary>
-        /// RSACryptoServiceProvider to manipulate to encrypt the information
-        /// </summary>
-        private RSACryptoServiceProvider _cryptoProvider;
-        public RSACryptoServiceProvider CryptoProvider
-        {
-            get { return _cryptoProvider; }
-        }
 
         /// <summary>
         /// Domain element constructor
@@ -37,69 +25,10 @@ namespace ConfigurationSettings
             return Domain;
         }
 
-        /// <summary>
-        /// Create the RSACryptoServiceProvider for the domain
-        /// </summary>
-        /// <param name="basePath">Path of the private key to open</param>
-        /// <returns></returns>
-        public bool InitElement(string basePath)
+        public string PrivateKeyPathAbsolute(string basePath)
         {
-            string path = Path.IsPathRooted(PrivateKeyFile) ? PrivateKeyFile : Path.Combine(basePath, @"keys\" + PrivateKeyFile);
-
-            if (String.IsNullOrEmpty(path) || !File.Exists(path))
-            {
-                throw new FileNotFoundException("The private key for domain " + Domain + " wasn't found: " + path);
-            }
-
-            try
-            {
-                _cryptoProvider = RSACryptoHelper.GetProviderFromKeyFile(path);
-            }
-            catch (Exception e)
-            {
-                throw new CryptographicException("Couldn't load the key '" + path + "' for domain " + Domain + ". Error message: " + e.Message);
-            }
-            if (_cryptoProvider == null)
-            {
-                throw new RSACryptoHelperException("Couldn't load the key '" + path + "' for domain " + Domain + ". Invalid key format or broken file");
-            }
-
-            return true;
+            return Path.IsPathRooted(PrivateKeyFile) ? PrivateKeyFile : Path.Combine(basePath, @"keys\" + PrivateKeyFile);
         }
 
-        /// <summary>
-        /// Finalizes an instance of the <see cref="DkimSigner"/> class.
-        /// </summary>
-        ~DomainElement()
-        {
-            Dispose(false);
-        }
-
-        /// <summary>
-        /// Performs application-defined tasks associated with freeing,
-        /// releasing, or resetting unmanaged resources.
-        /// </summary>
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        /// <summary>
-        /// Releases unmanaged and - optionally - managed resources
-        /// </summary>
-        /// <param name="disposing"><c>true</c> to release both managed and unmanaged resources;
-        /// <c>false</c> to release only unmanaged resources.</param>
-        private void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                if (_cryptoProvider != null)
-                {
-                    _cryptoProvider.Clear();
-                    _cryptoProvider = null;
-                }
-            }
-        }
     }
 }